library: hexbright/includes moved to hexbright, renamed to pc_stubs, read_adc
